+++ .github/workflows/bump-examples.yml
@@ -1,7 +1,7 @@
 name: Bump Examples
 on:
   schedule:
-    - chron: 0 0 0 * * 3
+    - cron: '0 0 * * 4'
 jobs:
   bump:
     runs-on: ubuntu-latest
@@ -11,14 +11,14 @@ jobs:
         with:
           node-version: lts/*
       - name: bump examples
-        run: node ./.github/bump-examples.js
+        run: node ./.github/scripts/bump-examples.js
       - name: submit pull request
         uses: peter-evans/create-pull-request@v3
         with:
           base: master
-          branch: bump-examples
-          commit_message: "chore(deps): update calcite components version"
+          branch: action/bump-examples
+          commit-message: "build(deps): update calcite components version"
           token: ${{ secrets.GITHUB_TOKEN }}
-          title: "chore(deps): update calcite components version"
-          body: This PR was automatically generated by the bump-examples GitHub action
-          labels: "chore"
+          title: "build(deps): update calcite components version"
+          body: This PR was automatically generated by the bump-examples GitHub Action
+          delete-branch: true
